invoice added cart commit

+++ lesson1/src/main/java/seminar2/FinanceDocument.java
@@ -6,7 +6,7 @@ package seminar2;
 *
 * */
 public interface FinanceDocument {
-    public  Float sumDoc();
-    public  Float vatDoc();
+    public  Double sumDoc();
+    public  Double vatDoc();
 
 }
